camera setting add

+++ src/trace/hit/hit_cylinder.c
@@ -6,7 +6,7 @@
 /*   By: sooyokim <sooyokim@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/08 09:26:15 by sooyokim          #+#    #+#             */
-/*   Updated: 2022/11/08 10:21:26 by sooyokim         ###   ########.fr       */
+/*   Updated: 2022/11/08 13:04:48 by sooyokim         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,8 +19,8 @@ int	hit_cy_cap(t_object *cy_obj, t_ray *ray, t_hit_record *rec, double height)
 	t_cylinder	*cy;
 	double		r;
 	t_vec3		center;
-	float		root; // float쓸지 double 쓸지 파악 필요
-	float		diameter;
+	double		root; // float쓸지 double 쓸지 파악 필요, 원래 float였음
+	double		diameter;
 
 	cy = cy_obj->element;
 	r = cy->diameter / 2;
